requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
huggingface-hub
